// src/app/components/navigation/navActions.enum.ts
/**
 *
 */
export enum NavActionsEnum {
  showStartingPage = 'showStartingPage',
  menuUsers = 'menuUsers',
  menuAssets = 'menuAssets',
  menuUnits = 'menuUnits',
  menuCategories = 'menuCategories',
  menuLists = 'menuLists',

  createUser = 'createUser',
  usersList = 'usersList',
  deleteSelectedUsers = 'deleteSelectedUsers',
  editSelectedUsers = 'editSelectedUsers',

  assetsList = 'assetsList',

  createUnit = 'createUnit',
  unitsList = 'unitsList',
  collapseUnits = 'collapseUnits',
  expandUnits = 'expandUnits',

  categoriesList = 'categoriesList',
  createCategory = 'createCategory',

  editCategoryColumns = 'editCategoryColumns',

  createAsset = 'createAsset',

  lists = 'lists',
  showWorkingList = 'showWorkingList',

  showMultiEdit = 'showMultiEdit',

  transferProtocol = 'transferProtocol',

  menuLocations = 'menuLocations',

  menuHistory = 'menuHistory'


}